requirements.txt
# ── Core ─────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.0.0
Pillow>=10.0.0

# ── Device control ───────────────────────────────
uiautomator2>=3.0.0
adbutils>=2.0.0

# ── LLM (OpenAI-compatible Kimi) ─────────────────
openai>=1.50.0

# ── Agent orchestration ──────────────────────────
langgraph>=0.2.0
langchain-core>=0.3.0

# ── UI ───────────────────────────────────────────
gradio>=4.40.0

# ── OCR (可选,M0 不强制装,装失败先靠 UI 树+VLM) ──
# paddlepaddle>=2.6.0
# paddleocr>=2.7.0

# sqlite3 / typing 等是 Python 内置,无需写
